+++ plugins/idp-basic/src/main/java/org/apache/gravitino/idp/storage/po/IdpUserPO.java
@@ -16,18 +16,27 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.gravitino.idp.basic.storage.relational.mapper.provider;
+package org.apache.gravitino.idp.storage.po;
 
-import com.google.common.collect.ImmutableList;
-import java.util.List;
-import org.apache.gravitino.idp.basic.storage.relational.mapper.IdpUserMetaMapper;
-import org.apache.gravitino.storage.relational.mapper.provider.MapperPackageProvider;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.ToString;
 
-/** Supplies built-in IdP mapper classes from the idp-basic plugin. */
-public class IdpBasicMapperPackageProvider implements MapperPackageProvider {
-
-  @Override
-  public List<Class<?>> getMapperClasses() {
-    return ImmutableList.of(IdpUserMetaMapper.class);
-  }
+@Getter
+@EqualsAndHashCode
+@ToString
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
+@Builder(setterPrefix = "with")
+public class IdpUserPO {
+  private Long userId;
+  private String userName;
+  private String passwordHash;
+  private Long currentVersion;
+  private Long lastVersion;
+  private Long deletedAt;
 }
